// src/app/core/services/model/tblMaquinas.ts
/**
 * Mi API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TblZona } from './tblZona';
import { TblListadoEquiposRuta } from './tblListadoEquiposRuta';
import { TblTipoMaquinas } from './tblTipoMaquinas';


export interface TblMaquinas { 
    idMaquina?: string | null;
    idTipo?: number | null;
    idZona?: number | null;
    nombre?: string | null;
    personalRequeridoMaquina?: number | null;
    personalRequeridoEmbalaje?: number | null;
    estado?: string | null;
    fechaCreacion?: string | null;
    equipoRuta?: Array<TblListadoEquiposRuta> | null;
    zona?: TblZona;
    tipoMaquina?: TblTipoMaquinas;
}