// src/app/core/services/model/tblFamiliaProductos.ts
/**
 * Mi API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TblCambiosCiclosIntermedios } from './tblCambiosCiclosIntermedios';


export interface TblFamiliaProductos { 
    idFamilia?: number;
    nombre?: string | null;
    estado?: boolean | null;
    tiempoFormulacion?: number | null;
    fechaCreacion?: string | null;
    cambiosCiclosIntermediosFamiliaProceso?: Array<TblCambiosCiclosIntermedios> | null;
    cambiosCiclosIntermediosFamiliaCambio?: Array<TblCambiosCiclosIntermedios> | null;
}